Update form validation examples
