+++ packages/react/src/examples/forms/validation.scss
@@ -14,6 +14,10 @@
     margin: 0 0 var(--spacing-l) 0;
   }
 
+  &__error-summary {
+    margin-bottom: var(--spacing-m);
+  }
+
   &__section {
     margin-bottom: var(--spacing-xl);
