+++ packages/react/src/examples/forms/validationUtils.ts
@@ -1,0 +1,123 @@
+/**
+ * Simple date validation function.
+ * Returns true if given date string is a valid date in DD.MM.YYYY format.
+ */
+export const isValidDate = (value?: string) => {
+  if (typeof value === 'string' && value.match(/^(\d{1,2}.){2}\d{4}$/)) {
+    const dateParts = value.split('.');
+    return !Number.isNaN(Date.parse(`${dateParts[1]}/${dateParts[0]}/${dateParts[2]}`));
+  }
+  return false;
+};
+
+const cities = [
+  'Akaa',
+  'Alajärvi',
+  'Alavus',
+  'Espoo',
+  'Forssa',
+  'Haapajärvi',
+  'Haapavesi',
+  'Hamina',
+  'Hanko',
+  'Harjavalta',
+  'Heinola',
+  'Helsinki',
+  'Huittinen',
+  'Hyvinkää',
+  'Hämeenlinna',
+  'Iisalmi',
+  'Ikaalinen',
+  'Imatra',
+  'Pietarsaari',
+  'Joensuu',
+  'Jyväskylä',
+  'Jämsä',
+  'Järvenpää',
+  'Kaarina',
+  'Kajaani',
+  'Kalajoki',
+  'Kangasala',
+  'Kankaanpää',
+  'Kannus',
+  'Karkkila',
+  'Kaskinen',
+  'Kauhajoki',
+  'Kauhava',
+  'Kauniainen',
+  'Kemi',
+  'Kemijärvi',
+  'Kerava',
+  'Keuruu',
+  'Kitee',
+  'Kiuruvesi',
+  'Kokemäki',
+  'Kokkola',
+  'Kotka',
+  'Kouvola',
+  'Kristiinankaupunki',
+  'Kuhmo',
+  'Kuopio',
+  'Kurikka',
+  'Kuusamo',
+  'Lahti',
+  'Laitila',
+  'Lappeenranta',
+  'Lapua',
+  'Lieksa',
+  'Lohja',
+  'Loimaa',
+  'Loviisa',
+  'Maarianhamina',
+  'Mikkeli',
+  'Mänttä',
+  'Naantali',
+  'Nivala',
+  'Nokia',
+  'Nurmes',
+  'Uusikaarlepyy',
+  'Närpiö',
+  'Orimattila',
+  'Orivesi',
+  'Oulainen',
+  'Oulu',
+  'Outokumpu',
+  'Paimio',
+  'Parainen',
+  'Parkano',
+  'Pieksämäki',
+  'Pori',
+  'Porvoo',
+  'Pudasjärvi',
+  'Pyhäjärvi',
+  'Raahe',
+  'Raasepori',
+  'Raisio',
+  'Rauma',
+  'Riihimäki',
+  'Rovaniemi',
+  'Saarijärvi',
+  'Salo',
+  'Sastamala',
+  'Savonlinna',
+  'Seinäjoki',
+  'Somero',
+  'Suonenjoki',
+  'Tampere',
+  'Tornio',
+  'Turku',
+  'Ulvila',
+  'Uusikaupunki',
+  'Vaasa',
+  'Valkeakoski',
+  'Vantaa',
+  'Varkaus',
+  'Viitasaari',
+  'Virrat',
+  'Ylivieska',
+  'Ylöjärvi',
+  'Ähtäri',
+  'Äänekoski',
+];
+
+export const getCitites = () => cities.map((city) => ({ label: city }));
